moved globals.d.ts and database.d.ts to types folder
